[bug-fix] Use float64 when converting np.ndarray to torch.tensor, cap Torch version to 1.7.x (#4610)
* Use float64 in GAIL tests

* Use float32 when converting np arrays by default

* Enforce torch 1.7.x or below

* Add comment about Windows install

* Adjust tests

+++ ml-agents/mlagents/trainers/tests/torch/test_reward_providers/utils.py
@@ -9,10 +9,12 @@ def create_agent_buffer(
 ) -> AgentBuffer:
     buffer = AgentBuffer()
     curr_observations = [
-        np.random.normal(size=shape) for shape in behavior_spec.observation_shapes
+        np.random.normal(size=shape).astype(np.float32)
+        for shape in behavior_spec.observation_shapes
     ]
     next_observations = [
-        np.random.normal(size=shape) for shape in behavior_spec.observation_shapes
+        np.random.normal(size=shape).astype(np.float32)
+        for shape in behavior_spec.observation_shapes
     ]
     action = behavior_spec.action_spec.random_action(1)[0, :]
     for _ in range(number):
